Move header & footer to components folder
